renaming project to RopeTeam
