+++ src/main/java/com/rei/ropeteam/EventPublisher.java
@@ -1,4 +1,4 @@
-package io.github.flock;
+package com.rei.ropeteam;
 
 public interface EventPublisher {
     void publish(Object event);
