+++ src/main/java/com/rei/ropeteam/EventSubscriber.java
@@ -1,4 +1,4 @@
-package io.github.flock;
+package com.rei.ropeteam;
 
 public interface EventSubscriber {
     void receiveEvent(Object event);
