+++ src/test/java/com/rei/ropeteam/spring/BlankEvent.java
@@ -1,0 +1,4 @@
+package com.rei.ropeteam.spring;
+
+public class BlankEvent {
+}
