--- src/test/java/io/github/flock/spring/BlankEvent.java
@@ -1,4 +1,0 @@
-package io.github.flock.spring;
-
-public class BlankEvent {
-}
